create DonationPage to manage sub-components inside
